+++ frontend/scstyle_01.css
@@ -1,6 +1,6 @@
 body {
     background-color: #c0c0c0;
-    font-family: "Gill Sans", sans-serif;
+    font-family: "Gill Sans", "Calabri", sans-serif;
 }
 
 h1 {
@@ -11,11 +11,18 @@ div {
     border-color: black;
     border-radius: 6px;
     max-width: 700px;
+    line-height: 40px;
+    margin: auto;    
 }
-
+div.noborder {
+    border-style:none;
+    text-align:center;    
+}
 div.sidebar {
+    padding:20px;
     float:left;
-    min-height:300px;
+    min-height:100px;
+    min-width:100px;
     background-color: #ffffff;
     max-width: 200px;
     text-align:center;
